📄 requirements.txt
gradio
torch
transformers
huggingface_hub
python-dotenv